chore: Update npm dependencies for aws-cdk and aws-sdk
This commit updates the npm dependencies for aws-cdk and aws-sdk to their latest versions. Specifically, it updates "@aws-cdk/aws-lambda" from "^1.204.0" to "^1.12.0" and "@aws-cdk/aws-s3" from "^1.204.0" to "^1.36.1". These updates ensure that the project is using the most up-to-date versions of these dependencies.

Note: This commit is based on the recent user commits and repository commits provided.